requirements.txt
llama-cpp-python
langchain
langchain_community
gradio